Removed diffignore file and two scripts.
This commit removes the .diffignore file, create-symlinks.sh script, and prepare-commit-msg script. The .diffignore file was deleted, which previously ignored certain types of files in the Git repository. The create-symlinks.sh script, which created symbolic links for prepare-commit-msg and .diffignore files, was also deleted. Lastly, the prepare-commit-msg script was removed, which utilized the OpenAI API to automatically generate commit messages based on the changes in the commit's diff. The script checked for additional commit message input, read the diff file, tokenized the changes, used the OpenAI API to generate a commit message, and saved it into a file. The prepare-commit-msg script provided a convenient way to automatically generate commit messages in a standardized format. However, it has been decided to remove these files and scripts from the repository.